+++ components/organisms/ReportAProblem.js
@@ -5,6 +5,7 @@ import { useTranslation } from "next-i18next";
 import { OptionalTextField } from "../molecules/OptionalTextField";
 import { Details } from "../molecules/Details";
 import { ActionButton } from "../atoms/ActionButton";
+import { ErrorLabel } from "../atoms/ErrorLabel";
 
 /**
  * Report a problem button to report technical issues on the page.
@@ -13,18 +14,18 @@ export function ReportAProblem(props) {
   const [submitted, setSubmitted] = useState(false);
   const { t, i18n } = useTranslation();
 
-  //Checking if at least one checkbox is selected
-  const [isCheckBoxSelected, setIsCheckBoxSelected] = useState(false);
+  const [submittedOnce, setSubmittedOnce] = useState(false);
 
   let onSubmitHandler = (e) => {
+    //Checking if at least one checkbox is selected
     let checkBoxSelected = false;
+
     //Check the checkboxes
     let inputElements = document.getElementsByTagName("input");
     for (let index = 0; index < inputElements.length; index++) {
       if (inputElements[index].type == "checkbox") {
         if (inputElements[index].checked) {
           checkBoxSelected = true;
-          setIsCheckBoxSelected(true);
         }
       }
     }
@@ -82,6 +83,8 @@ export function ReportAProblem(props) {
     if (checkBoxSelected) {
       setSubmitted(true);
     }
+    //Make sure the form was submitted at least once
+    setSubmittedOnce(true);
   };
 
   return (
@@ -156,6 +159,9 @@ export function ReportAProblem(props) {
             />
             <fieldset>
               <legend className="text-base sm:text-p font-body font-normal mb-6">
+                <b className="text-error-border-red mr-2" aria-hidden="true">
+                  *
+                </b>
                 {t("reportAProblemCheckAllThatApply", { lng: props.language })}
               </legend>
               <OptionalTextField
@@ -339,8 +345,10 @@ export function ReportAProblem(props) {
                 checkBoxStyle="mb-4"
                 controlValue={t("reportAProblemOther", { lng: props.language })}
               />
-              {isCheckBoxSelected ? (
-                <ErrorLabel message={"hghjg"} />
+              {submittedOnce ? (
+                <ErrorLabel
+                  message={t("reportAProblemError", { lng: props.language })}
+                />
               ) : undefined}
             </fieldset>
             <ActionButton